Fix chapter loading, player overlay, and modernize overflow menu
- Fetch chapters separately in detail view (server's single audiobook
  endpoint doesn't include them), fixing missing chapter info in
  progress box
- Replace .confirmationDialog overflow menu with modern bottom sheet
  featuring colored icons, subtitles, and rounded pill backgrounds
- Fix Chapter.duration type from Int? to Double? to match server JSON
- Hybrid speed picker with fine-tune +/- controls and preset grid
- Player uses overlay with drag-to-dismiss instead of .sheet/.fullScreenCover
- Chapters button always visible (disabled when no chapters loaded)
- Fix mini player time flicker by removing pulse animation

+++ Sappho/Domain/Model/Models.swift
@@ -247,7 +247,7 @@ struct Chapter: Codable, Identifiable {
     let chapterNumber: Int
     let filePath: String?
     let startTime: Double
-    let duration: Int?
+    let duration: Double?
     let fileSize: Int?
     let title: String?
     let createdAt: String?
@@ -269,7 +269,7 @@ struct Chapter: Codable, Identifiable {
         chapterNumber = try container.decodeIfPresent(Int.self, forKey: .chapterNumber) ?? 0
         filePath = try container.decodeIfPresent(String.self, forKey: .filePath)
         startTime = try container.decodeIfPresent(Double.self, forKey: .startTime) ?? 0
-        duration = try container.decodeIfPresent(Int.self, forKey: .duration)
+        duration = try container.decodeIfPresent(Double.self, forKey: .duration)
         fileSize = try container.decodeIfPresent(Int.self, forKey: .fileSize)
         title = try container.decodeIfPresent(String.self, forKey: .title)
         createdAt = try container.decodeIfPresent(String.self, forKey: .createdAt)
